\name{hscale_segment}
\alias{hscale_segment}
\title{Add scale bar to the ggplot}
\usage{
  hscale_segment(breaks, ...)
}
\arguments{
  \item{breaks}{Provide the dataframe in breaks}

  \item{...}{In this section the user can provide any
  arguments that will be used in geom_text option}
}
\description{
  This is function to convert scale bar with text in ggplot
  It requires the users to provide the dataframe "breaks"
}

