\name{my.minor.tick}
\alias{my.minor.tick}
\title{Add minor ticks in base graphics}
\usage{
  my.minor.tick(nx = 2, ny = 2, tick.ratio = 0.5,
    side = NA)
}
\arguments{
  \item{nx}{nx is the number of ticks in x direction}

  \item{ny}{ny is the number of ticks in y direction}

  \item{tick.ratio}{tick.ratio is the ratio of the ticks}

  \item{side}{specify the side where you want to add the
  minor ticks}
}
\description{
  This is function to add minor ticks in the base graphics
}

