\name{readts_EE}
\alias{readts_EE}
\title{Read time series output from EE}
\usage{
  readts_EE(textfile)
}
\arguments{
  \item{textfile}{textfile refers to the file used for the
  analysis. EFDC outputs usually have *.dat and *.txt
  format}
}
\description{
  This is function to read the time series output from EFDC
  model It will create a dataframe of three variables
}

