\name{condtosal}
\alias{condtosal}
\title{Convert the conductivity to salinity}
\usage{
  condtosal(cond, T, P)
}
\arguments{
  \item{cond}{Provide the conductivity values in mmho/cm}

  \item{T}{Provide the temperature in degree Celsius}

  \item{P}{P refers to pressure in decibars}
}
\description{
  Convert the conductivity to salinity
}
\examples{
condtosal(20,25,1)
}

